+++ src/geonodeobject.py
@@ -1,9 +1,9 @@
-from typing import List, Dict, Union
+from typing import List, Dict
 import json
 
 from src.geonodetypes import GeonodeCmdOutObjectKey, GeonodeCmdOutListKey
 from src.rest import GeonodeRest
-from src.cmdprint import show_list
+from src.cmdprint import print_list_on_cmd, print_json
 
 
 class GeonodeObjectHandler(GeonodeRest):
@@ -22,9 +22,9 @@ class GeonodeObjectHandler(GeonodeRest):
         """show list of geonode obj on the cmdline"""
         obj = self.list(**kwargs)
         if kwargs["json"]:
-            self.print_json(obj)
+            print_json(obj)
         else:
-            self.__class__.print_list_on_cmd(obj)
+            print_list_on_cmd(obj, self.LIST_CMDOUT_HEADER)
 
     def list(self, **kwargs) -> Dict:
         """returns dict of datasets from geonode
@@ -48,7 +48,7 @@ class GeonodeObjectHandler(GeonodeRest):
 
     def cmd_patch(self, pk: int, fields: str, **kwargs):
         obj = self.patch(pk, fields, **kwargs)
-        self.print_json(obj)
+        print_json(obj)
 
     def patch(self, pk: int, fields: str, **kwargs) -> Dict:
         """tries to generate  object from incoming json string
@@ -76,7 +76,7 @@ class GeonodeObjectHandler(GeonodeRest):
 
     def cmd_describe(self, pk: int, **kwargs):
         obj = self.get(pk=pk, **kwargs)
-        self.print_json(obj)
+        print_json(obj)
 
     def get(self, pk: int, **kwargs) -> Dict:
         """get details for a given pk
@@ -91,41 +91,3 @@ class GeonodeObjectHandler(GeonodeRest):
             endpoint=f"{self.ENDPOINT_NAME}/{pk}?page_size={kwargs['page_size']}"
         )
         return r[self.SINGULAR_RESOURCE_NAME]
-
-    @classmethod
-    def cmd_list_header(self) -> List[str]:
-        """returns the default header to print list on cmd
-
-        Returns:
-            List[str]: list of header elements as str
-        """
-        return [str(cmdoutkey) for cmdoutkey in self.LIST_CMDOUT_HEADER]
-
-    @classmethod
-    def print_list_on_cmd(cls, obj: Dict):
-        """print a beautiful list on the cmdline
-
-        Args:
-            obj (Dict): dict object to print on cmd line
-        """
-
-        def generate_line(i, obj: Dict, headers: List[GeonodeCmdOutObjectKey]) -> List:
-            return [cmdoutkey.get_key(obj[i]) for cmdoutkey in headers]
-
-        values = [
-            generate_line(i, obj, cls.LIST_CMDOUT_HEADER) for i in range(len(obj))
-        ]
-        show_list(headers=cls.cmd_list_header(), values=values)
-
-    @classmethod
-    def print_json(cls, json_str: Union[str, dict]):
-        """
-        Print the given JSON string or dictionary with an indentation of 2 spaces.
-
-        Args:
-            json_str (Union[str, dict]): The JSON string or dictionary to be printed.
-
-        Returns:
-            None
-        """
-        print(json.dumps(json_str, indent=2))
